Prevalence Rate data of Malaria in Africa

Data containing information on malaria prevalence in Africa. Data was extracted using the malariaAtlas package. If using this data, please cite:

data(malaria_africa_ts)

Format

An object of class tbl_ts (inherits from tbl_df, tbl, data.frame) with 1046 rows and 15 columns.

Details

Pfeffer, D.A., Lucas, T.C., May, D., Harris, J., Rozier, J., Twohig, K.A., Dalrymple, U., Guerra, C.A., Moyes, C.L., Thorn, M., Nguyen, M., et al. 2018. malariaAtlas: an R interface to global malariometric data hosted by the Malaria Atlas Project. Malaria Journal, 17(1), p.352.

Format: a 'tsibble' data frame with 1,117 observations on the following 15 variables.

who_region

A character of WHO region, the largest region size.

who_subregion

A character of WHO subregion, the second largest region size.

country

A character of countries, the smallest region size.

country_id

A three letter character code adhering to the standard, ISO 3166-1 alpha-3. For more detail see https://en.wikipedia.org/wiki/ISO_3166-1_alpha-3.

continent_id

A character of continent name

year

integer of year data recorded

month

integer of month data recorded

date

date class of year+month recorded. Defaults to day = 1 as day was not provided.

lower_age

lower age range

upper_age

upper age range

examined

integer, number of people examined

positive

numeric, number of positive cases

pr

positive rate - given from positive/examined

species

Mosquito species

method

method of measurement